chore: added flag interface and fixed st array tests

+++ model/client/account/queue_data.go
@@ -5,8 +5,8 @@ import "github.com/CreatureDev/xrpl-go/model/transactions/types"
 type QueueData struct {
 	TxnCount           uint64                  `json:"txn_count"`
 	AuthChangeQueued   bool                    `json:"auth_change_queued,omitempty"`
-	LowestSequence     uint64                  `json:"lowest_sequence,omitempty"`
-	HighestSequence    uint64                  `json:"highest_sequence,omitempty"`
+	LowestSequence     uint32                  `json:"lowest_sequence,omitempty"`
+	HighestSequence    uint32                  `json:"highest_sequence,omitempty"`
 	MaxSpendDropsTotal types.XRPCurrencyAmount `json:"max_spend_drops_total,omitempty"`
 	Transactions       []QueueTransaction      `json:"transactions,omitempty"`
 }
